@@ -1,10 +1,10 @@
 module github.com/goal-web/supports
 
-go 1.19
+go 1.25.0
 
 require (
 	github.com/apex/log v1.9.0
-	github.com/goal-web/contracts v0.1.66
+	github.com/goal-web/contracts v0.5.4
 	github.com/modood/table v0.0.0-20200225102042-88de94bb9876
 	github.com/pkg/errors v0.8.1
 	github.com/stretchr/testify v1.7.0
@@ -16,4 +16,3 @@ require (
 	github.com/smartystreets/goconvey v1.7.2 // indirect
 	gopkg.in/yaml.v3 v3.0.0-20200605160147-a5ece683394c // indirect
 )
-
