@@ -2,6 +2,7 @@ package utils
 
 import (
 	"errors"
+	"fmt"
 	"github.com/goal-web/contracts"
 	"reflect"
 	"strings"
@@ -13,6 +14,18 @@ func OnlyFields(fields contracts.Fields, keys ...string) contracts.Fields {
 
 	for _, key := range keys {
 		results[key] = fields[key]
+	}
+
+	return results
+}
+
+func FieldKeys[T any](fields map[string]T) []string {
+	var results = make([]string, len(fields))
+
+	i := 0
+	for key := range fields {
+		results[i] = key
+		i++
 	}
 
 	return results
@@ -63,7 +76,7 @@ func MergeFields(fields contracts.Fields, finalFields contracts.Fields) {
 	}
 }
 
-// GetStringField 获取 Fields 中的字符串，会尝试转换类型
+// GetStringField 获取 ToFields 中的字符串，会尝试转换类型
 func GetStringField(fields contracts.Fields, key string, defaultValues ...string) string {
 	if value, existsString := fields[key]; existsString {
 		if str, isString := value.(string); isString {
@@ -73,7 +86,7 @@ func GetStringField(fields contracts.Fields, key string, defaultValues ...string
 	return StringOr(defaultValues...)
 }
 
-// GetSubField 获取下级 Fields ，如果没有的话，匹配同前缀的放到下级 Fields 中
+// GetSubField 获取下级 ToFields ，如果没有的话，匹配同前缀的放到下级 ToFields 中
 func GetSubField(fields contracts.Fields, key string, defaultValues ...contracts.Fields) contracts.Fields {
 
 	if subField, isField := fields[key].(contracts.Fields); isField {
@@ -100,7 +113,7 @@ func GetSubField(fields contracts.Fields, key string, defaultValues ...contracts
 	return subField
 }
 
-// GetInt64Field 获取 Fields 中的 int64，会尝试转换类型
+// GetInt64Field 获取 ToFields 中的 int64，会尝试转换类型
 func GetInt64Field(fields contracts.Fields, key string, defaultValues ...int64) int64 {
 	var defaultValue int64 = 0
 	if len(defaultValues) > 0 {
@@ -116,7 +129,7 @@ func GetInt64Field(fields contracts.Fields, key string, defaultValues ...int64) 
 	}
 }
 
-// GetIntField 获取 Fields 中的 int，会尝试转换类型
+// GetIntField 获取 ToFields 中的 int，会尝试转换类型
 func GetIntField(fields contracts.Fields, key string, defaultValues ...int) int {
 	var defaultValue = 0
 	if len(defaultValues) > 0 {
@@ -132,7 +145,7 @@ func GetIntField(fields contracts.Fields, key string, defaultValues ...int) int 
 	}
 }
 
-// GetFloatField 获取 Fields 中的 float32，会尝试转换类型
+// GetFloatField 获取 ToFields 中的 float32，会尝试转换类型
 func GetFloatField(fields contracts.Fields, key string, defaultValues ...float32) float32 {
 	var defaultValue float32 = 0
 	if len(defaultValues) > 0 {
@@ -148,7 +161,7 @@ func GetFloatField(fields contracts.Fields, key string, defaultValues ...float32
 	}
 }
 
-// GetFloat64Field 获取 Fields 中的 float64，会尝试转换类型
+// GetFloat64Field 获取 ToFields 中的 float64，会尝试转换类型
 func GetFloat64Field(fields contracts.Fields, key string, defaultValues ...float64) float64 {
 	var defaultValue float64 = 0
 	if len(defaultValues) > 0 {
@@ -164,7 +177,7 @@ func GetFloat64Field(fields contracts.Fields, key string, defaultValues ...float
 	}
 }
 
-// GetBoolField 获取 Fields 中的 bool，会尝试转换类型
+// GetBoolField 获取 ToFields 中的 bool，会尝试转换类型
 func GetBoolField(fields contracts.Fields, key string, defaultValues ...bool) bool {
 	var defaultValue = false
 	if len(defaultValues) > 0 {
@@ -176,15 +189,21 @@ func GetBoolField(fields contracts.Fields, key string, defaultValues ...bool) bo
 	return defaultValue
 }
 
-// ToFields 尝试把一个变量转换成 Fields 类型
+// ToFields 尝试把一个变量转换成 ToFields 类型
 func ToFields(anyValue any) (contracts.Fields, error) {
 	fields := contracts.Fields{}
 	switch paramValue := anyValue.(type) {
+	case contracts.FieldsProvider:
+		fields = paramValue.ToFields()
 	case contracts.Fields:
 		fields = paramValue
 	case map[string]any:
 		for key, value := range paramValue {
 			fields[key] = value
+		}
+	case map[any]any:
+		for key, value := range paramValue {
+			fields[fmt.Sprintf("%v", key)] = value
 		}
 	case map[string]int:
 		for key, value := range paramValue {
@@ -244,8 +263,17 @@ func ToFields(anyValue any) (contracts.Fields, error) {
 		}
 	default:
 		paramType := reflect.ValueOf(anyValue)
-
 		switch paramType.Kind() {
+		case reflect.Ptr: // 结构体指针
+			if paramType.Elem().Kind() == reflect.Struct {
+				EachStructField(paramType.Elem(), paramType.Elem().Interface(), func(field reflect.StructField, value reflect.Value) {
+					if field.IsExported() {
+						fields[SnakeString(field.Name)] = value.Interface()
+					} else {
+						fields[SnakeString(field.Name)] = nil
+					}
+				})
+			}
 		case reflect.Struct: // 结构体
 			EachStructField(paramType, anyValue, func(field reflect.StructField, value reflect.Value) {
 				if field.IsExported() {
@@ -260,7 +288,7 @@ func ToFields(anyValue any) (contracts.Fields, error) {
 				fields[name] = paramType.MapIndex(key).Interface()
 			}
 		default:
-			return nil, errors.New("不支持转 contracts.Fields 的类型： " + paramType.String())
+			return nil, errors.New("不支持转 contracts.ToFields 的类型： " + paramType.String())
 		}
 	}
 	return fields, nil
