@@ -47,6 +47,16 @@ func StringOr(otherStr ...string) string {
 	return ""
 }
 
+// IntOr 尽量不返回0值
+func IntOr(otherInt ...int) int {
+	for _, i := range otherInt {
+		if i != 0 {
+			return i
+		}
+	}
+	return 0
+}
+
 // SnakeString 蛇形字符串
 func SnakeString(s string) string {
 	data := make([]byte, 0, len(s)*2)
@@ -77,10 +87,10 @@ func CamelString(s string) string {
 	num := len(s) - 1
 	for i := 0; i <= num; i++ {
 		d := s[i]
-		if k == false && d >= 'A' && d <= 'Z' {
+		if !k && d >= 'A' && d <= 'Z' {
 			k = true
 		}
-		if d >= 'a' && d <= 'z' && (j || k == false) {
+		if d >= 'a' && d <= 'z' && (j || !k) {
 			d = d - 32
 			j = false
 			k = true
